--- lib/src/widgets/movie_detail_button.dart
@@ -1,49 +1,0 @@
-import 'package:flutter/material.dart';
-import '../resources/movie_strings.dart';
-import '../../utils/text_styles.dart';
-import '../../utils/constants.dart';
-
-class MovieDetailButton extends StatelessWidget {
-  MovieDetailButton();
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.symmetric(
-        horizontal: Constants.paddingButton,
-      ),
-      child: ElevatedButton(
-        style: ButtonStyle(
-          backgroundColor: MaterialStateProperty.resolveWith(
-            (
-              states,
-            ) =>
-                Colors.black,
-          ),
-        ),
-        onPressed: () {
-          Navigator.of(context).pop();
-        },
-        child: Row(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            Padding(
-              padding: const EdgeInsets.all(
-                Constants.paddingButtonChildren,
-              ),
-              child: Icon(
-                Icons.arrow_back_sharp,
-                size: Constants.paddingButtonIconSize,
-                color: Colors.white,
-              ),
-            ),
-            Text(
-              MovieStrings.movieDetailBackText,
-              style: TextStyles.styleButton,
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
